@@ -6,79 +6,69 @@
   "author": "Kostya <holypower77@gmail.com>",
   "license": "UNLICENSED",
   "private": true,
+  "engines": {
+    "node": ">=18.0.0"
+  },
   "scripts": {
-    "start:app": "yarn --cwd packages/app start",
+    "start:tgc": "yarn --cwd packages/tgc start",
     "clean-dist": "rm -rf .storybook/dist && mkdir .storybook/dist",
+    "clean:nodemodules": "rm -rf node_modules && rm -rf packages/tgc/node_modules",
     "lint:js": "eslint --ext .js,.jsx,.ts,.tsx packages",
     "lint:js:fix": "yarn run lint:js --fix",
     "lint:css": "stylelint 'packages/**/*.scss'",
     "lint:css:fix": "yarn run lint:css --fix",
     "test": "jest --passWithNoTests --env=jsdom",
-    "storybook:build": "yarn clean-dist && build-storybook -o .storybook/dist -c .storybook",
-    "storybook:start": "start-storybook -p 9000",
+    "storybook:build": "yarn clean-dist && storybook build -o .storybook/dist -c .storybook",
+    "storybook:start": "storybook dev -p 9000",
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/eslint-parser": "^7.15.8",
-    "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.16.7",
-    "@storybook/addon-essentials": "^6.5.10",
-    "@storybook/addon-storyshots": "^6.5.10",
-    "@storybook/addon-storyshots-puppeteer": "^6.5.10",
-    "@storybook/addon-toolbars": "^6.5.14",
-    "@storybook/builder-webpack5": "^6.5.14",
-    "@storybook/manager-webpack5": "^6.5.14",
+    "@babel/core": "^7.20.0",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
+    "@storybook/addon-essentials": "^7.0.0-alpha.49",
     "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^6.5.14",
-    "@storybook/testing-react": "^1.3.0",
+    "@storybook/react": "^7.0.0-alpha.49",
+    "@storybook/react-webpack5": "^7.0.0-alpha.49",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.0.4",
-    "@types/jest": "^27.4.0",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
-    "babel-jest": "^29.0.1",
-    "babel-loader": "^8.2.3",
-    "babel-plugin-transform-runtime": "^6.23.0",
+    "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.2.4",
+    "@types/wouter": "^2.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.46.1",
+    "babel-loader": "^9.1.0",
     "crypto-browserify": "^3.12.0",
-    "css-loader": "^6.4.0",
-    "eslint": "^8.1.0",
-    "eslint-loader": "^4.0.2",
+    "css-loader": "^6.7.3",
+    "eslint": "^8.30.0",
     "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import-helpers": "^1.1.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "file-loader": "^6.2.0",
-    "gulp": "^4.0.2",
+    "eslint-plugin-import-helpers": "^1.3.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.5.0",
-    "husky": "^7.0.0",
+    "husky": "^8.0.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-image-snapshot": "^5.1.1",
-    "lerna": "^4.0.0",
-    "lint-staged": "^12.3.3",
-    "mini-css-extract-plugin": "^2.6.0",
+    "jest": "^29.3.1",
+    "lerna": "^6.1.0",
+    "lint-staged": "^13.1.0",
+    "mini-css-extract-plugin": "^2.7.2",
+    "postcss-scss": "^4.0.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "playwright": "^1.25.1",
-    "puppeteer": "^17.0.0",
-    "react-intl": "^5.21.0",
-    "react-test-renderer": "^18.2.0",
-    "sass-loader": "^12.2.0",
-    "sass-variable-loader": "^0.1.2",
+    "sass": "^1.57.0",
+    "sass-loader": "^13.2.0",
+    "storybook": "^7.0.0-alpha.49",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.1",
-    "stylelint": "^13.13.1",
-    "stylelint-order": "^4.1.0",
-    "stylelint-scss": "^4.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.4.4",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.59.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.3.1"
+    "stylelint": "^14.16.0",
+    "stylelint-order": "^5.0.0",
+    "stylelint-scss": "^4.3.0",
+    "typescript": "^4.9.4",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1"
   },
   "jest": {
     "moduleNameMapper": {
