@@ -11,47 +11,38 @@
     "build:analyze": "webpack-bundle-analyzer webpack.analyze.json"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.7.2",
+    "@reduxjs/toolkit": "^1.9.1",
     "b_": "^1.3.4",
-    "classnames": "^2.3.1",
-    "dank-twitch-irc": "^4.3.0",
-    "firebase": "^9.6.10",
-    "flagsmith": "^2.0.9",
+    "classnames": "^2.3.2",
+    "flagsmith": "^3.14.3",
     "jws": "^4.0.0",
-    "react-charty": "^0.1.7",
     "react-helmet": "^6.1.0",
-    "react-hot-loader": "^4.13.0",
-    "react-intl": "^5.20.13",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-spring": "^9.4.4",
-    "react-swipeable-views": "^0.14.0",
+    "react-intl": "^6.2.5",
+    "react-redux": "^8.0.5",
+    "react-spring": "^9.6.1",
     "react-virtualized": "^9.22.3",
-    "react-window": "^1.8.6",
-    "react-window-infinite-loader": "^1.0.7",
-    "tmi.js": "^1.8.5"
+    "react-window": "^1.8.8",
+    "react-window-infinite-loader": "^1.0.8"
   },
   "devDependencies": {
     "@types/b_": "^1.3.2",
-    "@types/classnames": "^2.3.1",
     "@types/jws": "^3.2.4",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@types/react-helmet": "^6.1.6",
-    "@types/react-intl": "^3.0.0",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/react-swipeable-views": "^0.13.1",
     "@types/react-virtualized": "^9.21.21",
-    "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window-infinite-loader": "^1.0.6",
-    "@types/tmi.js": "^1.8.2",
-    "compression-webpack-plugin": "^9.2.0",
-    "dotenv": "^16.0.0",
-    "webpack": "^5.53.0",
-    "webpack-bundle-analyzer": "^4.5.0"
+    "compression-webpack-plugin": "^10.0.0",
+    "dotenv": "^16.0.3",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0"
   },
   "peerDependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "wouter": "^2.9.0"
+  },
+  "resolutions": {
+    "react-helmet/react-side-effect": "2.1.0"
   }
 }
