@@ -1,17 +1,16 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react';
 import React from 'react';
 
 import { StoryOverview } from '../../containers/stories';
-import { STORIES } from '../constants';
 
 import Logo from './logo';
 
 export default {
-    title: `${STORIES.pcCommon}/Logo`,
+    title: 'Platform Components/Common/Logo',
     component: Logo,
-} as ComponentMeta<typeof Logo>;
+} as Meta<typeof Logo>;
 
-export const Overview: ComponentStory<typeof Logo> = (args) => (
+export const Overview: StoryFn<typeof Logo> = (args) => (
     <StoryOverview componentName="Logo">
         <Logo {...args} />
     </StoryOverview>
