@@ -4,7 +4,7 @@ import { Spin } from 'platform-components';
 import React, { lazy, Suspense } from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider as ReduxProvider } from 'react-redux';
-import { Routes, Route, BrowserRouter, Navigate } from 'react-router-dom';
+import { Route, Switch } from 'wouter';
 
 import { withHeader, CommonFooter } from './common.components';
 import NoLimits from './components/no-limits/no-limits';
@@ -25,41 +25,41 @@ function createContainer(targetParent: Element) {
 const container = document.getElementById(rootContainerId) ?? createContainer(document.body);
 const root = createRoot(container);
 
-const LiveChat = lazy(() => import('./components/live-chat/live-chat'));
-const OverallStats = lazy(() => import('./components/overall-stats/overall-stats'));
-const Home = lazy(() => import('./components/home/home'));
-const UserHistory = lazy(() => import('./components/user-history/user-history'));
-const About = lazy(() => import('./components/about/about'));
+// const LiveChat = lazy(() => import('./components/live-chat/live-chat'));
+// const OverallStats = lazy(() => import('./components/overall-stats/overall-stats'));
+const Home = lazy(() => import('./pages/home/home'));
+// const UserHistory = lazy(() => import('./components/user-history/user-history'));
+const About = lazy(() => import('./pages/about/about'));
 
 const App = () => {
     return (
         <FlagsmithProvider
             flagsmith={flagsmith}
-            options={{ environmentID: process.env.FLAGSMITH_API_KEY || '' }}
+            options={{ 
+                environmentID: process.env.FLAGSMITH_API_KEY || '',
+                angularHttpClient: false,
+            }}
         >
             <ReduxProvider store={store}>
                 <LanguageProvider>
                     <UserProvider>
                         <NotificationProvider>
                             <Suspense fallback={<Spin center/>}>
-                                <BrowserRouter>
                                     <LocationProvider>
-                                        <Routes>
-                                            <Route path="/">
-                                                <Route element={<Home/>} index/>
-                                                <Route element={withHeader(<LiveChat/>)} path="live-chat"/>
+                                            <Switch>
+                                                <Route component={Home} path="/" />
+                                                <Route component={About} path="/about" />
+                                                {/* <Route element={withHeader(<LiveChat/>)} path="live-chat"/>
                                                 <Route element={withHeader(<UserHistory/>)} path="messages">
                                                     <Route element={withHeader(<UserHistory/>)} path=":username"/>
                                                 </Route>
                                                 <Route element={withHeader(<OverallStats/>)} path="overall-stats"/>
                                                 <Route element={withHeader(<NoLimits/>)} path="no-limits"/>
                                                 <Route element={withHeader(<About/>)} path="about"/>
-                                                <Route element={<Navigate to="/"/>} path="*"/>
-                                            </Route>
-                                        </Routes>
+                                                <Route element={<Navigate to="/"/>} path="*"/> */}
+                                            </Switch>
                                         <CommonFooter/>
                                     </LocationProvider>
-                                </BrowserRouter>
                             </Suspense>
                         </NotificationProvider>
                     </UserProvider>
